Renamed areThereEmojis and areThereSpaces file names.
