+++ frontend/jobapptrackerfrontend/src/__tests__/areThereEmojis.test.js
