+++ frontend/jobapptrackerfrontend/src/functions/emojiChecker/areThereEmojis.js
