+++ frontend/jobapptrackerfrontend/src/functions/spaceChecker/areThereSpaces.js
@@ -1,0 +1,9 @@
+
+/*
+areThereSpaces(input: String) : Bool
+
+checks if there is a space in input.
+*/
+export const areThereSpaces = (input) => {
+
+}
